Conflicts:
	src/server/game/Battlegrounds/Battleground.cpp
	src/server/game/World/World.cpp

+++ src/server/game/Entities/Corpse/Corpse.cpp
@@ -122,7 +122,7 @@ void Corpse::SaveToDB()
     stmt->setUInt32(index++, uint32(m_time));                                         // time
     stmt->setUInt8 (index++, GetType());                                              // corpseType
     stmt->setUInt32(index++, GetInstanceId());                                        // instanceId
-    stmt->setUInt16(index++, GetPhaseMask());                                         // phaseMask
+    stmt->setUInt32(index++, GetPhaseMask());                                         // phaseMask
     trans->Append(stmt);
 
     CharacterDatabase.CommitTransaction(trans);
@@ -186,7 +186,7 @@ bool Corpse::LoadCorpseFromDB(uint32 guid, Field* fields)
     m_time = time_t(fields[12].GetUInt32());
 
     uint32 instanceId  = fields[14].GetUInt32();
-    uint32 phaseMask   = fields[15].GetUInt16();
+    uint32 phaseMask   = fields[15].GetUInt32();
 
     // place
     SetLocationInstanceId(instanceId);
